backend_flask/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
cryptography
PyMySQL
python-dotenv
